Classe aluno corrigida

+++ Projeto1GenerationBrasil/src/cadastroUsuario/Aluno.java
@@ -5,16 +5,16 @@ import java.time.LocalDate;
 public class Aluno extends Pessoa {
 		
 	private int anoEscolar;
-	private int mediaGeral_anoAnterior;
-	private int mediaGeral_anoAtual;
-	private int mediaMatematica_anoAnterior;
-	private int mediaMatematica_anoAtual;
+	private double mediaGeral_anoAnterior;
+	private double mediaGeral_anoAtual;
+	private double mediaMatematica_anoAnterior;
+	private double mediaMatematica_anoAtual;
 	private String senha;
 	
 	public Aluno(int id, String nome, String sobrenomesDoMeio, String ultimoSobrenome, String rg,
 			LocalDate dataDeNascimento, String endereco, String celular, String email,String senha, int anoEscolar,
-			int mediaGeral_anoAnterior, int mediaGeral_anoAtual, int mediaMatematica_anoAnterior,
-			int mediaMatematica_anoAtual) {
+			double mediaGeral_anoAnterior, double mediaGeral_anoAtual, double mediaMatematica_anoAnterior,
+			double mediaMatematica_anoAtual) {
 		super(id, nome, sobrenomesDoMeio, ultimoSobrenome, rg, dataDeNascimento, endereco, celular, email);
 		this.anoEscolar = anoEscolar;
 		this.mediaGeral_anoAnterior = mediaGeral_anoAnterior;
@@ -40,7 +40,7 @@ public class Aluno extends Pessoa {
 		this.anoEscolar = anoEscolar;
 	}
 
-	public int getMediaGeral_anoAnterior() {
+	public double getMediaGeral_anoAnterior() {
 		return mediaGeral_anoAnterior;
 	}
 
@@ -48,7 +48,7 @@ public class Aluno extends Pessoa {
 		this.mediaGeral_anoAnterior = mediaGeral_anoAnterior;
 	}
 
-	public int getMediaGeral_anoAtual() {
+	public double getMediaGeral_anoAtual() {
 		return mediaGeral_anoAtual;
 	}
 
@@ -56,7 +56,7 @@ public class Aluno extends Pessoa {
 		this.mediaGeral_anoAtual = mediaGeral_anoAtual;
 	}
 
-	public int getMediaMatematica_anoAnterior() {
+	public double getMediaMatematica_anoAnterior() {
 		return mediaMatematica_anoAnterior;
 	}
 
@@ -64,7 +64,7 @@ public class Aluno extends Pessoa {
 		this.mediaMatematica_anoAnterior = mediaMatematica_anoAnterior;
 	}
 
-	public int getMediaMatematica_anoAtual() {
+	public double getMediaMatematica_anoAtual() {
 		return mediaMatematica_anoAtual;
 	}
